+++ sample.project.clj
@@ -55,4 +55,11 @@
                  ;; directories. By default every directory on the
                  ;; classpath is refreshed. Value is passed through to clojure.tools.namespace.repl/set-refresh-dirs
                  ;; https://github.com/clojure/tools.namespace/blob/f3f5b29689c2bda53b4977cf97f5588f82c9bd00/src/main/clojure/clojure/tools/namespace/repl.clj#L164
-                 :refresh-dirs ["src" "test"]})
+                 :refresh-dirs ["src" "test"]
+
+
+                 ;; Use this flag to specify your own flag to add to
+                 ;; cause test-refresh to focus. Intended to be used
+                 ;; to let you specify a shorter flag than the default
+                 ;; :test-refresh/focus.
+                 :focus-flag :test-refresh/focus})